fix warning C4244 on Windows
'initializing': conversion from 'uint64_t' to 'int',  possible loss of data

+++ Python/optimizer_bytecodes.c
@@ -846,7 +846,7 @@ dummy_func(void) {
             ctx->done = true;
             break;
         }
-        int returning_stacklevel = this_instr->operand1;
+        int returning_stacklevel = (int)this_instr->operand1;
         if (ctx->curr_frame_depth >= 2) {
             PyCodeObject *expected_code = ctx->frames[ctx->curr_frame_depth - 2].code;
             if (expected_code == returning_code) {
@@ -872,7 +872,7 @@ dummy_func(void) {
             break;
         }
         _Py_BloomFilter_Add(dependencies, returning_code);
-        int returning_stacklevel = this_instr->operand1;
+        int returning_stacklevel = (int)this_instr->operand1;
         if (frame_pop(ctx, returning_code, returning_stacklevel)) {
             break;
         }
@@ -892,7 +892,7 @@ dummy_func(void) {
             break;
         }
         _Py_BloomFilter_Add(dependencies, returning_code);
-        int returning_stacklevel = this_instr->operand1;
+        int returning_stacklevel = (int)this_instr->operand1;
         if (frame_pop(ctx, returning_code, returning_stacklevel)) {
             break;
         }
